+++ packages/delegation-deployments/src/index.ts
@@ -27,6 +27,8 @@ export const CHAIN_ID = {
   celo: 0xa4ec,
   ronin: 0x7e4,
   tempoMainnet: 0x1079,
+  citreaMainnet: 0x1012,
+  mantleMainnet: 0x1388,
   // Testnets
   bscTestnet: 0x61,
   arbitrumSepolia: 0x66eee,
@@ -48,6 +50,7 @@ export const CHAIN_ID = {
   celoSepolia: 0xaa044c,
   roninSaigon: 0x31769,
   tempoModeratoTestnet: 0xa5bf,
+  mantleSepolia: 0x138b,
   // decommissioned
   lineaGoerli: 0xe704,
 };
@@ -111,6 +114,8 @@ export const DELEGATOR_CONTRACTS: DeployedContracts = {
     [CHAIN_ID.celo]: DEPLOYMENTS_1_3_0,
     [CHAIN_ID.ronin]: DEPLOYMENTS_1_3_0,
     [CHAIN_ID.tempoMainnet]: DEPLOYMENTS_1_3_0,
+    [CHAIN_ID.citreaMainnet]: DEPLOYMENTS_1_3_0,
+    [CHAIN_ID.mantleMainnet]: DEPLOYMENTS_1_3_0,
     // Testnets
     [CHAIN_ID.bscTestnet]: DEPLOYMENTS_1_3_0,
     [CHAIN_ID.citreaTestnet]: DEPLOYMENTS_1_3_0,
@@ -132,5 +137,6 @@ export const DELEGATOR_CONTRACTS: DeployedContracts = {
     [CHAIN_ID.celoSepolia]: DEPLOYMENTS_1_3_0,
     [CHAIN_ID.roninSaigon]: DEPLOYMENTS_1_3_0,
     [CHAIN_ID.tempoModeratoTestnet]: DEPLOYMENTS_1_3_0,
+    [CHAIN_ID.mantleSepolia]: DEPLOYMENTS_1_3_0,
   },
 };