+++ packages/smart-accounts-kit/vitest.config.ts
@@ -4,5 +4,9 @@ export default defineConfig({
   test: {
     globals: true,
     environment: 'node',
+    env: {
+      // Don't track analytics during unit tests.
+      DO_NOT_TRACK: 'true',
+    },
   },
 });