+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [24.0.0]
+
+## [23.0.0]
+
 ## [22.0.0]
 
 ## [21.0.0]
@@ -39,7 +43,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [7.0.0]
 
-[Unreleased]: https://github.com/metamask/smart-accounts-kit/compare/delegator-sdk-monorepo@22.0.0...HEAD
+[Unreleased]: https://github.com/metamask/smart-accounts-kit/compare/delegator-sdk-monorepo@24.0.0...HEAD
+[24.0.0]: https://github.com/metamask/smart-accounts-kit/compare/delegator-sdk-monorepo@23.0.0...delegator-sdk-monorepo@24.0.0
+[23.0.0]: https://github.com/metamask/smart-accounts-kit/compare/delegator-sdk-monorepo@22.0.0...delegator-sdk-monorepo@23.0.0
 [22.0.0]: https://github.com/metamask/smart-accounts-kit/compare/delegator-sdk-monorepo@21.0.0...delegator-sdk-monorepo@22.0.0
 [21.0.0]: https://github.com/metamask/smart-accounts-kit/compare/delegator-sdk-monorepo@20.0.0...delegator-sdk-monorepo@21.0.0
 [20.0.0]: https://github.com/metamask/smart-accounts-kit/compare/delegator-sdk-monorepo@19.0.0...delegator-sdk-monorepo@20.0.0